+++ src/main/java/com/esri/core/geometry/OperatorSimpleRelation.java
@@ -43,14 +43,14 @@ public abstract class OperatorSimpleRelation extends Operator {
                                     SpatialReference sr,
                                     ProgressTracker progressTracker);
 
-    public HashMap<Integer, Boolean> execute(Geometry inputGeom1,
+    public HashMap<Long, Boolean> execute(Geometry inputGeom1,
                                              GeometryCursor geometryCursor2,
                                              SpatialReference sr,
                                              ProgressTracker progressTracker) {
-        HashMap<Integer, Boolean> hashMap = new HashMap<>();
+        HashMap<Long, Boolean> hashMap = new HashMap<>();
         Geometry inputGeom2;
         while ((inputGeom2 = geometryCursor2.next()) != null) {
-            int index = geometryCursor2.getGeometryID();
+            Long index = geometryCursor2.getGeometryID();
             if ((progressTracker != null) && !(progressTracker.progress(-1, -1)))
                 throw new RuntimeException("user_canceled");
             hashMap.put(index, execute(inputGeom1, inputGeom2, sr, progressTracker));
