+++ src/main/java/com/esri/core/geometry/StringCursor.java
@@ -5,7 +5,7 @@ import java.util.Iterator;
 public abstract class StringCursor implements Iterator<String> {
     public abstract String next();
 
-    public abstract int getID();
+    public abstract long getID();
 
     public abstract boolean hasNext();
 }
